requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
graphql-core==3.2.3
idna==3.4
psycopg2-binary==2.9.5
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
strawberry-graphql==0.152.0
strawberry-graphql-django==0.8.2
typing_extensions==4.4.0
urllib3==1.26.14
